ref(plugins): Inline PluginComponentBase into its two subclasses
The abstract PluginComponentBase class was only extended by
PluginSettings and IssueActions. Inline all base class logic
(state management, save/load lifecycle, renderField, API client)
directly into each consumer and delete the base class file.
